pyproject.toml
[project]
name = "cryptonomicsai"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "curl-cffi>=0.15.0",
    "firecrawl-py>=2.9.0",
    "openai>=1.88.0",
    "pydantic>=2.11.7",
    "requests>=2.32.4",
    "selectolax>=0.4.7",
]
